added own service for mute button in home component
